refactor(utils): improve organization and separation of concerns
Move VS Code-specific utilities from utils/ to frontend/:
- dialogs.ts → frontend/ui/dialogs.ts
- commandUtils.ts → frontend/system/commandUtils.ts
- activeFileGuards.ts → frontend/editor/activeFileGuards.ts

Consolidate root-level utils files into appropriate subdirectories:
- helpers.ts → utils/core/async.ts (sleep, sleepWithAbort)
- urlUtils.ts → utils/core/urlCore.ts (normalizeUrl)
- pendingStateManager.ts → common/state/pendingStateManager.ts
- settingsQueries.ts → utils/config/settingsQueries.ts

Remove unused utility functions:
- generateTripleBackslashFixes, generateBackslashFixesWithCase,
  generateEnvironmentSpacingFixes, generateDecoratedGreekShortcuts,
  generateTextCommandNormalization from replacement/helpers.ts
- readFileBytesSync, filterExistingFiles from workspaceFS.ts

This improves maintainability by:
- Enforcing proper layering (VS Code APIs in frontend/, pure utils in utils/)
- Consolidating scattered utilities into organized modules
- Removing dead code that adds maintenance burden
