+++ src/commands/files/fileSelectionCommands.ts
@@ -8,7 +8,7 @@ import { fileLister } from '@frontend/files';
 import { showInfoMessage } from '@frontend/ui/messageUtils';
 import * as logger from '@logger/logUtils';
 import { WorkspaceFS } from '@utils/files';
-import { selectFile, selectFiles } from '@utils/dialogs';
+import { selectFile, selectFiles } from '@frontend/ui/dialogs';
 
 const CHANNEL = 'fileSelectionCommands';
 logger.initialize(CHANNEL);
