+++ src/commands/system/xmlCommands.ts
@@ -10,7 +10,7 @@ import * as logger from '@logger/logUtils';
 import {
   getActiveEditorWithGuards,
   logGuardFailure,
-} from '@utils/editor/activeFileGuards';
+} from '@frontend/editor/activeFileGuards';
 
 const CHANNEL = 'XmlCommands';
 logger.initialize(CHANNEL);
