+++ src/common/state/pendingStateManager.ts
