+++ src/frontend/editor/activeFileGuards.ts
