+++ src/frontend/index.ts
@@ -7,6 +7,7 @@ export * from './latex/linter';
 export * from './ui/messageUtils';
 export * from './ui/instruction';
 export * from './ui/diffView';
+export * from './ui/dialogs';
 export * from './webview/html';
 export * from './terminal';
 export * from './agents/AgentDirectoryManager';
@@ -15,3 +16,5 @@ export * from './secretManager';
 export * from './setup';
 export * from './media/img';
 export * from './media/audio';
+export * from './system/commandUtils';
+export * from './editor/activeFileGuards';
