+++ src/utils/config/index.ts
@@ -1,3 +1,4 @@
 export * from './configUtils';
 export * from './configConversion';
 export * from './constants';
+export * from './settingsQueries';
