+++ src/utils/config/settingsQueries.ts
@@ -1,3 +1,7 @@
+/**
+ * VS Code settings query constants for extension configuration.
+ */
+
 export const SETTINGS_QUERY = {
   EXTENSION: '@ext:texra-ai.texra',
   WORKFLOW_AGENTS: '@ext:texra-ai.texra texra.agents',
