+++ src/utils/core/index.ts
@@ -1,5 +1,5 @@
 /**
- * Core utilities - consolidated path, string, and type helpers.
+ * Core utilities - consolidated path, string, type, async, and URL helpers.
  */
 
 export { toPosixPath, getPathSegments, normalizeLatexPath } from './pathCore';
@@ -11,3 +11,5 @@ export {
   type SerializedError,
 } from './stringCore';
 export { contentToString, isObject, type MessageContent } from './typeGuards';
+export { sleep, sleepWithAbort } from './async';
+export { normalizeUrl } from './urlCore';
