+++ src/utils/system/index.ts
@@ -1,4 +1,3 @@
 export * from './execUtils';
-export * from './commandUtils';
 export * from './toolUtils';
 export * from './platformPaths';
