+++ src/webview/MainViewMessageHandler.ts
@@ -13,12 +13,9 @@ import { agentDirectories } from '@frontend/agents';
 import { showInstructionWithSuppress } from '@frontend/ui/instruction';
 import { computeModelOptions } from '@model/computeModelOptions';
 import { getConfig, setConfig } from '@utils/config';
-import {
-  safeExecuteCommand,
-  checkCoreDependencies,
-  getToolDocsCommand,
-} from '@utils/system';
-import { SETTINGS_QUERY } from '@utils/settingsQueries';
+import { safeExecuteCommand } from '@frontend/system';
+import { checkCoreDependencies, getToolDocsCommand } from '@utils/system';
+import { SETTINGS_QUERY } from '@utils/config';
 import { AUTH_COMMANDS, getAuthStatus } from '@commands/auth';
 import { PROVIDER_URLS } from '@commands/api/apiKeyCommands';
